backend/requirements.txt
Flask==2.3.2
RPi.GPIO==0.7.0
pygcode==0.7.4